+++ tests/Paket.Tests/Simplifier/BasicScenarioSpecs.fs
@@ -41,7 +41,7 @@ let projects1 = [
     ReferencesFile.FromLines [|"B";"C"|] ] |> List.zip [dummyProjectFile; dummyProjectFile]
 
 [<Test>]
-let ``should remove one level deep indirect dependencies from dep and ref files``() = 
+let ``should remove one level deep transitive dependencies from dep and ref files``() = 
     let before = PaketEnv.create dummyDir depFile1 lockFile1 projects1
     
     match Simplifier.simplify false before with
@@ -84,7 +84,7 @@ let projects2 = [
     ReferencesFile.FromLines [|"C";"D";"E"|] ] |> List.zip [dummyProjectFile; dummyProjectFile]
 
 [<Test>]
-let ``should remove all indirect dependencies from dep file recursively``() =
+let ``should remove all transitive dependencies from dep file recursively``() =
     let before = PaketEnv.create dummyDir depFile2 lockFile2 projects2
     
     match Simplifier.simplify false before with